Logic of usePath hook has been moved to hooks folder
